backend/servicio-pacientes/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic[email]==2.5.0
python-dotenv==1.0.0
requests==2.31.0
